+++ Battelle.EPA.WideAreaDecon.Model/CharacterizationSampling/IPhaseLagCalculator.cs
@@ -2,6 +2,6 @@
 {
     public interface IPhaseLagCalculator
     {
-        public double CalculatePhaseLagTime(double numberLabs, double sampleTimeTransmitted, double surfaceAreaToBeWiped, double surfaceAreaToBeHepa);
+        public double CalculatePhaseLagTime(int numberLabs, double sampleTimeTransmitted, double surfaceAreaToBeWiped, double surfaceAreaToBeHepa);
     }
 }
