+++ Battelle.EPA.WideAreaDecon.Model/IncidentCommand/ISuppliesCostCalculator.cs
@@ -3,6 +3,6 @@
     public interface ISuppliesCostCalculator
     {
         public double CalculateSuppliesCost(double _numberTeams, double surfaceAreaToBeSourceReduced, double personnelRoundTripDays, double _surfaceAreaToBeHepa
-            , double _surfaceAreaToBeWiped, double numberLabs, double sampleTimeTransmitted);
+            , double _surfaceAreaToBeWiped, int numberLabs, double sampleTimeTransmitted);
     }
 }
